centralize most bot messages
