+++ server/plugin.go
@@ -176,21 +176,8 @@ func (p *Plugin) OnDisconnectedTokenHandler(userID string) {
 		p.API.LogWarn("Unable clean invalid token for the user", "user_id", userID, "error", err2.Error())
 		return
 	}
-	channel, appErr := p.API.GetDirectChannel(userID, p.GetBotUserID())
-	if appErr != nil {
-		p.API.LogWarn("Unable to get direct channel for send message to user", "user_id", userID, "error", appErr.Error())
-		return
-	}
 
-	connectURL := p.GetURL() + "/connect"
-	_, appErr = p.API.CreatePost(&model.Post{
-		UserId:    p.GetBotUserID(),
-		ChannelId: channel.Id,
-		Message:   "Your connection to Microsoft Teams has been lost. " + fmt.Sprintf("[Click here to reconnect your account](%s).", connectURL),
-	})
-	if appErr != nil {
-		p.API.LogWarn("Unable to send direct message to user", "user_id", userID, "error", appErr.Error())
-	}
+	p.notifyUserDisconnected(userID)
 }
 
 func (p *Plugin) GetClientForUser(userID string) (msteams.Client, error) {
